Put BaseTimelineEventDataPOarser inside its own file. Added exception for if an event in the xml file has 2 default event contents. Added eventnumber to default content exception.
