+++ Assets/_Core/Scripts/Internal/Timeline/Xml/BaseTimelineEventDataParser.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 3efcd6eb0bc706c4f9964e4681deeddd
+timeCreated: 1546975796
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
